requirements.txt
pymongo==3.12.0
dnspython==2.1.0
fastapi==0.67.0
